adding version numbers to setup.py and moving the predict function to utils_with_wndcharm.py as it contains a reference to wndcharm
