+++ setup.py
@@ -8,5 +8,11 @@ setup(
     long_description=open('README.md').read(),
     author='Duke Lungmap Team',
     description='A machine learning pipeline used to predict anatomy from Lungmap images.',
-    install_requires=['numpy', 'opencv-python', 'matplotlib', 'scipy', 'pandas', 'scikit-learn'],
+    install_requires=[
+        'numpy==1.13.1',
+        'opencv-python==3.2.0.7',
+        'matplotlib==2.0.2',
+        'scipy==0.19.1',
+        'pandas==0.19.2',
+        'scikit-learn==0.18.2'],
 )
